added sever side check to deleteMessage to ensure the delete only valid for owner of card

+++ app/models/message.server.ts
@@ -1,9 +1,19 @@
 import type { message } from "@prisma/client";
 import { prisma } from "~/db.server";
+import { requireUserId } from "~/session.server";
 
 export type { message } from "@prisma/client";
 
-export function deleteMessage({ message_id }: Pick<message, "message_id">) {
+export async function deleteMessage({ cardOwnerId, request, message_id }: Pick<message, "message_id"> & {
+  cardOwnerId: number;
+  request: Request;
+}) {
+  const userId = await requireUserId(request);
+  if (userId !== cardOwnerId) {
+    // TODO: throw error
+    return
+  }
+
   return prisma.message.deleteMany({
     where: { message_id },
   });
